requirements.txt
# =============================================================================
# Google Ads API Agent — Python Dependencies
# =============================================================================

# --- Core: Anthropic API ---
anthropic>=0.43.0

# --- Google Ads API ---
google-ads>=28.1.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# --- Web Server (for API deployment) ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# --- Cloudinary (Creative Tools) ---
cloudinary>=1.40.0

# --- Utilities ---
python-dotenv>=1.0.0
httpx>=0.27.0
requests>=2.31.0

# --- Optional: CLI enhancements ---
rich>=13.7.0
